Modified timing code.

+++ emulator/globals.h
@@ -28,6 +28,6 @@ extern unsigned short EX;
 extern unsigned long cycles_counter;
 extern unsigned long emu_freq;
 extern unsigned int emu_speed;
-extern unsigned long nsec_per_chunk;
+extern unsigned int nsec_per_chunk;
 
 #endif /* GLOBALS_H_INCLUDED */
